Azure APP-Service backend and Xamarin forms frontend.
Acty-master is the master branch for the new azure app service backend.
The backend service project is 'Acty.Service.sln'. This contains all the
same API's to save and retrive data from documentDB, store media in the
blob storage etc.. Same goodness. Run the acty.Service and try out.

The solution Acty.Mobile.sln is still a work in progress. I have copied
the code of xamarin sample app Sport.Mobile and repurposed it for my
app. First I just wanted to try out the login experience and then
introduce a tabbed form to accomplish the task I want. Currently Iam
stuck in the login part it self. The app successfully calls the backend
and is able to get google credentials, but then fails to pass the and
actually make an API call to backend hosted as an azure API service. I
will work on it.